+++ src/app/admin/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import {
   Users,
   Mail,
@@ -10,58 +10,29 @@ import {
   RefreshCw,
   Pencil,
   Trash2,
-  X,
-  AlertTriangle,
 } from "lucide-react";
-
-interface AdminUser {
-  id: string;
-  email: string;
-  name: string;
-  avatarUrl?: string;
-  role: string;
-  birthDate?: string;
-  gender?: string;
-  createdAt: string;
-  lastSignInAt?: string;
-  provider: string;
-}
+import type { AdminUser } from "@/components/admin/types";
+import { useAdminUsers } from "@/components/admin/useAdminUsers";
+import { UserEditModal } from "@/components/admin/UserEditModal";
+import { UserDeleteModal } from "@/components/admin/UserDeleteModal";
 
 export default function AdminPage() {
-  const [users, setUsers] = useState<AdminUser[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+  const {
+    users,
+    loading,
+    error,
+    editLoading,
+    deleteLoading,
+    fetchUsers,
+    saveUser,
+    deleteUser,
+  } = useAdminUsers();
 
   // 編集モーダル
   const [editingUser, setEditingUser] = useState<AdminUser | null>(null);
-  const [editForm, setEditForm] = useState({ name: "", role: "" });
-  const [editLoading, setEditLoading] = useState(false);
 
   // 削除確認モーダル
   const [deletingUser, setDeletingUser] = useState<AdminUser | null>(null);
-  const [deleteLoading, setDeleteLoading] = useState(false);
-
-  const fetchUsers = async () => {
-    setLoading(true);
-    setError(null);
-
-    try {
-      const response = await fetch("/api/admin/users");
-      if (!response.ok) {
-        throw new Error("ユーザー情報の取得に失敗しました");
-      }
-      const data = await response.json();
-      setUsers(data.users);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "エラーが発生しました");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchUsers();
-  }, []);
 
   const formatDate = (dateString: string | undefined) => {
     if (!dateString) return "-";
@@ -82,69 +53,21 @@ export default function AdminPage() {
     return colors[provider] || "bg-gray-100 text-gray-700";
   };
 
-  // 編集開始
-  const handleEditClick = (user: AdminUser) => {
-    setEditingUser(user);
-    setEditForm({ name: user.name, role: user.role });
-  };
-
   // 編集保存
-  const handleEditSave = async () => {
+  const handleEditSave = async (form: { name: string; role: string }) => {
     if (!editingUser) return;
-
-    setEditLoading(true);
-    try {
-      const response = await fetch(`/api/admin/users/${editingUser.id}`, {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(editForm),
-      });
-
-      if (!response.ok) {
-        throw new Error("更新に失敗しました");
-      }
-
-      // ユーザー一覧を更新
-      setUsers((prev) =>
-        prev.map((u) =>
-          u.id === editingUser.id ? { ...u, ...editForm } : u
-        )
-      );
+    const ok = await saveUser(editingUser.id, form);
+    if (ok) {
       setEditingUser(null);
-    } catch (err) {
-      alert(err instanceof Error ? err.message : "エラーが発生しました");
-    } finally {
-      setEditLoading(false);
     }
-  };
-
-  // 削除確認
-  const handleDeleteClick = (user: AdminUser) => {
-    setDeletingUser(user);
   };
 
   // 削除実行
   const handleDeleteConfirm = async () => {
     if (!deletingUser) return;
-
-    setDeleteLoading(true);
-    try {
-      const response = await fetch(`/api/admin/users/${deletingUser.id}`, {
-        method: "DELETE",
-      });
-
-      if (!response.ok) {
-        const data = await response.json();
-        throw new Error(data.error || "削除に失敗しました");
-      }
-
-      // ユーザー一覧から削除
-      setUsers((prev) => prev.filter((u) => u.id !== deletingUser.id));
+    const ok = await deleteUser(deletingUser.id);
+    if (ok) {
       setDeletingUser(null);
-    } catch (err) {
-      alert(err instanceof Error ? err.message : "エラーが発生しました");
-    } finally {
-      setDeleteLoading(false);
     }
   };
 
@@ -303,14 +226,14 @@ export default function AdminPage() {
                     <td className="px-4 py-4">
                       <div className="flex items-center justify-end gap-2">
                         <button
-                          onClick={() => handleEditClick(user)}
+                          onClick={() => setEditingUser(user)}
                           className="p-2 text-gray-400 hover:text-blue-600 hover:bg-blue-50 rounded-lg transition-colors"
                           title="編集"
                         >
                           <Pencil className="w-4 h-4" />
                         </button>
                         <button
-                          onClick={() => handleDeleteClick(user)}
+                          onClick={() => setDeletingUser(user)}
                           className="p-2 text-gray-400 hover:text-red-600 hover:bg-red-50 rounded-lg transition-colors"
                           title="削除"
                         >
@@ -328,134 +251,22 @@ export default function AdminPage() {
 
       {/* 編集モーダル */}
       {editingUser && (
-        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
-          <div className="bg-white rounded-xl p-6 w-full max-w-md mx-4 shadow-xl">
-            <div className="flex items-center justify-between mb-4">
-              <h3 className="text-lg font-bold text-gray-800">ユーザー編集</h3>
-              <button
-                onClick={() => setEditingUser(null)}
-                className="p-1 hover:bg-gray-100 rounded-lg"
-              >
-                <X className="w-5 h-5 text-gray-500" />
-              </button>
-            </div>
-
-            <div className="space-y-4">
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  名前
-                </label>
-                <input
-                  type="text"
-                  value={editForm.name}
-                  onChange={(e) =>
-                    setEditForm((prev) => ({ ...prev, name: e.target.value }))
-                  }
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">
-                  役割
-                </label>
-                <select
-                  value={editForm.role}
-                  onChange={(e) =>
-                    setEditForm((prev) => ({ ...prev, role: e.target.value }))
-                  }
-                  className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
-                >
-                  <option value="editor">Editor</option>
-                  <option value="admin">Admin</option>
-                </select>
-              </div>
-
-              <div className="text-sm text-gray-500">
-                <p>メール: {editingUser.email}</p>
-                <p>ID: {editingUser.id}</p>
-              </div>
-            </div>
-
-            <div className="flex gap-3 mt-6">
-              <button
-                onClick={() => setEditingUser(null)}
-                className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-colors"
-              >
-                キャンセル
-              </button>
-              <button
-                onClick={handleEditSave}
-                disabled={editLoading}
-                className="flex-1 px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors disabled:opacity-50 flex items-center justify-center gap-2"
-              >
-                {editLoading && <Loader2 className="w-4 h-4 animate-spin" />}
-                保存
-              </button>
-            </div>
-          </div>
-        </div>
+        <UserEditModal
+          user={editingUser}
+          loading={editLoading}
+          onClose={() => setEditingUser(null)}
+          onSave={handleEditSave}
+        />
       )}
 
       {/* 削除確認モーダル */}
       {deletingUser && (
-        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
-          <div className="bg-white rounded-xl p-6 w-full max-w-md mx-4 shadow-xl">
-            <div className="flex items-center gap-3 mb-4">
-              <div className="w-10 h-10 bg-red-100 rounded-full flex items-center justify-center">
-                <AlertTriangle className="w-5 h-5 text-red-600" />
-              </div>
-              <h3 className="text-lg font-bold text-gray-800">
-                ユーザーを削除しますか？
-              </h3>
-            </div>
-
-            <div className="bg-gray-50 rounded-lg p-4 mb-4">
-              <div className="flex items-center gap-3">
-                {deletingUser.avatarUrl ? (
-                  <img
-                    src={deletingUser.avatarUrl}
-                    alt={deletingUser.name}
-                    className="w-12 h-12 rounded-full object-cover"
-                  />
-                ) : (
-                  <div className="w-12 h-12 rounded-full bg-gray-200 flex items-center justify-center">
-                    <span className="text-gray-500 font-medium text-lg">
-                      {deletingUser.name.charAt(0).toUpperCase()}
-                    </span>
-                  </div>
-                )}
-                <div>
-                  <p className="font-medium text-gray-800">
-                    {deletingUser.name}
-                  </p>
-                  <p className="text-sm text-gray-500">{deletingUser.email}</p>
-                </div>
-              </div>
-            </div>
-
-            <p className="text-sm text-red-600 mb-6">
-              この操作は取り消せません。ユーザーのデータは完全に削除されます。
-            </p>
-
-            <div className="flex gap-3">
-              <button
-                onClick={() => setDeletingUser(null)}
-                className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-colors"
-              >
-                キャンセル
-              </button>
-              <button
-                onClick={handleDeleteConfirm}
-                disabled={deleteLoading}
-                className="flex-1 px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700 transition-colors disabled:opacity-50 flex items-center justify-center gap-2"
-              >
-                {deleteLoading && <Loader2 className="w-4 h-4 animate-spin" />}
-                削除する
-              </button>
-            </div>
-          </div>
-        </div>
+        <UserDeleteModal
+          user={deletingUser}
+          loading={deleteLoading}
+          onClose={() => setDeletingUser(null)}
+          onConfirm={handleDeleteConfirm}
+        />
       )}
     </div>
   );
