+++ src/components/admin/UserDeleteModal.tsx
@@ -1,0 +1,76 @@
+"use client";
+
+import { AlertTriangle, Loader2 } from "lucide-react";
+import type { AdminUser } from "./types";
+
+interface UserDeleteModalProps {
+  user: AdminUser;
+  loading: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+}
+
+export function UserDeleteModal({
+  user,
+  loading,
+  onClose,
+  onConfirm,
+}: UserDeleteModalProps) {
+  return (
+    <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
+      <div className="bg-white rounded-xl p-6 w-full max-w-md mx-4 shadow-xl">
+        <div className="flex items-center gap-3 mb-4">
+          <div className="w-10 h-10 bg-red-100 rounded-full flex items-center justify-center">
+            <AlertTriangle className="w-5 h-5 text-red-600" />
+          </div>
+          <h3 className="text-lg font-bold text-gray-800">
+            ユーザーを削除しますか？
+          </h3>
+        </div>
+
+        <div className="bg-gray-50 rounded-lg p-4 mb-4">
+          <div className="flex items-center gap-3">
+            {user.avatarUrl ? (
+              <img
+                src={user.avatarUrl}
+                alt={user.name}
+                className="w-12 h-12 rounded-full object-cover"
+              />
+            ) : (
+              <div className="w-12 h-12 rounded-full bg-gray-200 flex items-center justify-center">
+                <span className="text-gray-500 font-medium text-lg">
+                  {user.name.charAt(0).toUpperCase()}
+                </span>
+              </div>
+            )}
+            <div>
+              <p className="font-medium text-gray-800">{user.name}</p>
+              <p className="text-sm text-gray-500">{user.email}</p>
+            </div>
+          </div>
+        </div>
+
+        <p className="text-sm text-red-600 mb-6">
+          この操作は取り消せません。ユーザーのデータは完全に削除されます。
+        </p>
+
+        <div className="flex gap-3">
+          <button
+            onClick={onClose}
+            className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-colors"
+          >
+            キャンセル
+          </button>
+          <button
+            onClick={onConfirm}
+            disabled={loading}
+            className="flex-1 px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700 transition-colors disabled:opacity-50 flex items-center justify-center gap-2"
+          >
+            {loading && <Loader2 className="w-4 h-4 animate-spin" />}
+            削除する
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
