+++ src/components/admin/UserEditModal.tsx
@@ -1,0 +1,91 @@
+"use client";
+
+import { useState } from "react";
+import { X, Loader2 } from "lucide-react";
+import type { AdminUser } from "./types";
+
+interface UserEditModalProps {
+  user: AdminUser;
+  loading: boolean;
+  onClose: () => void;
+  onSave: (form: { name: string; role: string }) => void;
+}
+
+export function UserEditModal({
+  user,
+  loading,
+  onClose,
+  onSave,
+}: UserEditModalProps) {
+  const [editForm, setEditForm] = useState({ name: user.name, role: user.role });
+
+  return (
+    <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
+      <div className="bg-white rounded-xl p-6 w-full max-w-md mx-4 shadow-xl">
+        <div className="flex items-center justify-between mb-4">
+          <h3 className="text-lg font-bold text-gray-800">ユーザー編集</h3>
+          <button
+            onClick={onClose}
+            className="p-1 hover:bg-gray-100 rounded-lg"
+          >
+            <X className="w-5 h-5 text-gray-500" />
+          </button>
+        </div>
+
+        <div className="space-y-4">
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              名前
+            </label>
+            <input
+              type="text"
+              value={editForm.name}
+              onChange={(e) =>
+                setEditForm((prev) => ({ ...prev, name: e.target.value }))
+              }
+              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              役割
+            </label>
+            <select
+              value={editForm.role}
+              onChange={(e) =>
+                setEditForm((prev) => ({ ...prev, role: e.target.value }))
+              }
+              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+            >
+              <option value="editor">Editor</option>
+              <option value="admin">Admin</option>
+            </select>
+          </div>
+
+          <div className="text-sm text-gray-500">
+            <p>メール: {user.email}</p>
+            <p>ID: {user.id}</p>
+          </div>
+        </div>
+
+        <div className="flex gap-3 mt-6">
+          <button
+            onClick={onClose}
+            className="flex-1 px-4 py-2 border border-gray-300 text-gray-700 rounded-lg hover:bg-gray-50 transition-colors"
+          >
+            キャンセル
+          </button>
+          <button
+            onClick={() => onSave(editForm)}
+            disabled={loading}
+            className="flex-1 px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-colors disabled:opacity-50 flex items-center justify-center gap-2"
+          >
+            {loading && <Loader2 className="w-4 h-4 animate-spin" />}
+            保存
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
