+++ src/components/admin/types.ts
@@ -1,0 +1,12 @@
+export interface AdminUser {
+  id: string;
+  email: string;
+  name: string;
+  avatarUrl?: string;
+  role: string;
+  birthDate?: string;
+  gender?: string;
+  createdAt: string;
+  lastSignInAt?: string;
+  provider: string;
+}
